requirements.txt
Pillow
discord-py-slash-command<=3.0.1
discord.py
urllib3
aiohttp
jishaku
psutil
asyncpg